Fixed the compare date in the CreateNote activity; now, if the end date selected is before the create date, a notification will always pop up. Added plateColor method to the CreateNote Java class that will add a random color to each note when created.
Updated the database version to 2.
Added a new column to the My Note table that takes an integer data type that represents a random color when the note is created.
Modified the addMyNote to take an additional parameter (color).

Removed the plateColor method and the import: android.graphics.Color from the NoteList adapter Java class.